# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: EasyConf Reviews Microservice
  description: "EasyConf is web-based conference management system, designed to organize\
    \ and administrate academic conferences, workshops, and events. It offers a platform\
    \ for authors, reviewers, and organizers (chairs) to handle paper submissions,\
    \ peer reviews, and program scheduling efficiently. It provides tools for managing\
    \ the review process, generating program schedules, and handling communication\
    \ between organizers and participants.\n\nThe Reviews Microservice handles the\
    \ process of reviewing papers before a conference. This microservice is responsible\
    \ for assigning papers to reviews, allowing reviewers to add their comments, and\
    \ allowing the PC chairs to finalize the decision about acceptance and rejection\
    \ of the submitted papers.\n\nThe different roles that the users can have are\
    \ specified in the Users Microservice and the process for submitting papers to\
    \ conferences is specified in the Submissions Microservice.\n\nTECHNICAL NOTE:\
    \ in every endpoint, we require the requesting user to provide their ID. I.e.,\
    \ the requesterID parameter must be provided. It is used for authentication in\
    \ order to check if the user is allowed to perform certain actions."
  version: 0.0.1-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /conferences/{conferenceID}/tracks/{trackID}/bidding-deadline:
    get:
      tags:
      - Conference Tracks
      summary: Get the bidding deadline
      description: "Responds with the bidding deadline for the given track. If the\
        \ deadline hasn't been set, it will automatically be set to a few days after\
        \ the submission deadline for the track. Once the bidding deadline passes,\
        \ the bidding phase of the track automatically ends and the review assignments\
        \ phase begins."
      operationId: getBiddingDeadline
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of a conference
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of a track
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful pull of the bidding deadline
          content:
            application/json:
              schema:
                type: string
                format: date-time
        "403":
          description: "Forbidden. The requester must be a chair of the track, or\
            \ a reviewer in the track."
        "404":
          description: Not Found. The requested track or conference were not found.
    put:
      tags:
      - Conference Tracks
      summary: Changes the bidding deadline
      description: "Changes the bidding deadline to the one provided in the body.\
        \ If the deadline isn't set, it will automatically be set to a few days after\
        \ the submission deadline for the track. Once the bidding deadline passes,\
        \ the bidding phase of the track automatically ends and the review assignments\
        \ phase begins."
      operationId: setBiddingDeadline
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of a conference
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of a track
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              description: The date for the deadline of the bid
              format: date-time
        required: true
      responses:
        "409":
          description: "Conflict. An invalid bidding deadline was given, or the old\
            \ bidding deadline has already passed. The new given bidding deadline,\
            \ must also not have passed already."
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. The requester must be a chair of the track.
        "200":
          description: Success.
        "404":
          description: Not Found. The requested track or conference were not found.
...
  /papers/{paperID}/reviews/phase:
    get:
      tags:
      - Reviews
      summary: Get the review phase of a paper
      description: Responds with how far along a paper is in the review process. Each
        paper whithin a track has its own phase. For more information check the description
        of the returned enum.
      operationId: getPhase
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: paperID
        in: path
        description: The ID of the paper
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: "Forbidden. The requester must be a chair of the track the\
            \ paper is in, or a reviewer assigned to that paper."
        "404":
          description: Not Found. The requested paper was not found.
        "200":
          description: Successful retrieval of the paper review phase
          content:
            application/json:
              schema:
                type: string
                description: "Exactly how far along a paper is in the review process\n\
                  \nBEFORE_REVIEW: Before the Review phase of the track the paper\
                  \ is in\n\nIN_REVIEW: The reviewers all need to submit a review.\
                  \ Once that happens the discussion phase for the paper automatically\
                  \ starts.\n\nIN_DISCUSSION: The reviewers can now see each others\
                  \ reviews and comment on them. If all of the reviews are either\
                  \ positive or negative, the chair of the track can finalize the\
                  \ reviews which automatically puts the paper into the REVIEW phase.\n\
                  \nREVIEWED: The reviews cannot be changed now. Once all of the papers\
                  \ in a track are REVIEWED, the track goes into the final phase and\
                  \ the reviews can be viewed by the authors."
                enum:
                - BEFORE_REVIEW
                - IN_REVIEW
                - IN_DISCUSSION
                - REVIEWED
...
  /conferences/{conferenceID}/tracks/{trackID}/phase:
    get:
      tags:
      - Conference Tracks
      summary: Get the review phase
      description: Responds with the review phase of the track. Consult with the documentation
        of the returned enum for more information.
      operationId: getPhase_1
      parameters:
      - name: requesterID
        in: query
        description: The ID of the user making the request.
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of the conference the track is in.
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of the track.
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "200":
          description: Successful pull of the review phase
          content:
            application/json:
              schema:
                type: string
                description: "Phases of the review process for a track.\n\nSUBMITTING:\
                  \ The review process hasn't started. Papers are still being submitted.\n\
                  \nBIDDING: Papers cannot be submitted now. Reviewers can bid on\
                  \ the papers, to indicate their expertise level. Ends after the\
                  \ bidding deadline for the track passes.\n\nASSIGNING: Chairs of\
                  \ the track can now assign papers to reviewers in the track.\n\n\
                  REVIEWING: Reviewing process. During this phase, the papers in the\
                  \ track are either in Review or in Discussion.\n\nFINAL: The reviews\
                  \ for all of the papers in the track have been finalized. They are\
                  \ now accessible to authors."
                enum:
                - SUBMITTING
                - BIDDING
                - ASSIGNING
                - REVIEWING
                - FINAL
        "403":
          description: "Forbidden. The requester must be a chair of the track, or\
            \ a reviewer in the track."
        "404":
          description: Not Found. The requested track or conference were not found.
  /conferences/{conferenceID}/tracks/{trackID}/papers:
    get:
      tags:
      - Conference Tracks
      summary: Get the papers for track
      description: Returns all the papers in the given track of a conference. You
        can request the full paper from another api point.
      operationId: getPapers
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of a conference
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of a track
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found. The requested track or conference was not found.
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. The requester must be a chair of the track or a
            reviewer in the track.
        "200":
          description: Successful retrieval of papers in given track of given conference
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaperSummaryWithID'
  /conferences/{conferenceID}/tracks/{trackID}/analytics:
    get:
      tags:
      - Conference Tracks
      summary: Get the summary stats of a track
      description: "Returns the numbers of accepted, rejected and not-yet-decided\
        \ papers."
      operationId: getAnalytics
      parameters:
      - name: requesterID
        in: query
        description: The ID of a user making the request
        required: true
        schema:
          type: integer
          format: int64
      - name: conferenceID
        in: path
        description: The ID of a conference
        required: true
        schema:
          type: integer
          format: int64
      - name: trackID
        in: path
        description: The ID of a track
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error. An unexpected server error occurred.
        "403":
          description: Forbidden. The requester has to be a chair of the track.
        "404":
          description: Not Found. The requested track was not found.
        "200":
          description: Successful retrieval of track summary stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackAnalytics'
components:
  schemas:
...
    PaperSummaryWithID:
      type: object
      properties:
        paperID:
          type: integer
          description: The ID of the paper
          format: int64
          example: 2
        title:
          type: string
          description: The title of the paper
          example: Tails of frogs
        abstractSection:
          type: string
          description: The abstract of the paper
          example: Some sample abstract
      description: "Includes the title, abstract and ID of a paper"
    TrackAnalytics:
      type: object
      properties:
        accepted:
          type: integer
          description: Number of accepted papers
          format: int32
          example: 0
        rejected:
          type: integer
          description: Number of rejected papers
          format: int32
          example: 1
        unknown:
          type: integer
          description: Number of papers that are not yet decided
          format: int32
          example: 2
      description: "Counts of accepted, rejected and unknown papers"